Lesson added  9

+++ Lesson9/Program.cs
@@ -1,6 +1,6 @@
 ﻿//Задайте значение N. Напишите программу, которая выведет все натуральные числа в промежутке от 1 до N.
 /*
-void ShowNumbers(double n)
+void ShowNumbers(int n)
 {
     if(n > 1) ShowNumbers(n - 1);
     Console.Write(n + " ");
@@ -10,7 +10,7 @@ ShowNumbers(5);
 */
 //Напишите программу, которая будет принимать на вход число и возвращать сумму его цифр.
 /*
-double SumOfDigits(double num)
+double SumOfDigits(int num)
 {
     if(num != 0) return SumOfDigits(num / 10) + num % 10;
     return 0;
@@ -20,7 +20,7 @@ Console.WriteLine(SumOfDigits(1234));
 */
 //Задайте значения M и N. Напишите программу, которая выведет все натуральные числа в промежутке от M до N.
 /*
-void ShowNumbers(double m, double n)
+void ShowNumbers(int m, int n)
 {
     if(m > n)
     {   ShowNumbers (m - 1, n);
